Another Headings and Paragraphs change (3)
I added another heading and paragraph change to the bottom of the page to make it suit "The Coffee Shop" theme and make the customer feel more inclined to come into the shop. I changed the I "promise" part to "our goal" and talked about serving the customers with only the best.